supabase/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
psycopg[binary]==3.3.3
python-dotenv==1.2.2
requests==2.32.3
openpyxl
apscheduler==3.10.4
